handle multiple tokenlogger items, improve structure for redesign
# Conflicts:
#	src/plugins/locales/fr_FR.js
