--- src/Steps/TokenLoggerDashboard.vue
@@ -1,190 +1,0 @@
-<template>
-  <div class="wrapper">
-    <div
-      v-if="loading"
-      class="status-message"
-    >
-      {{ status }}
-    </div>
-
-    <template v-for="activity in applet.activities">
-      <div
-        v-if="applet"
-        :key="activity._id"
-      >
-        <h4 class="activity-header">
-          {{ activity.question.en || activity.description.en }}
-        </h4>
-
-        <v-card
-          v-for="item in activity.items"
-          :key="item['_id']"
-          class="chart-card"
-        >
-          <header>
-            <h5>{{ item.description.en || item.label.en }}</h5>
-          </header>
-
-          <token-chart
-            plot-id="token-chart"
-            :minValue="item.minValue"
-            :maxValue="item.maxValue"
-            :data="item.responses"
-            :features="item.responseOptions"
-            :versionsByDate="item.dateToVersions"
-            :versions="applet.versions"
-            :timezone="item.timezoneStr"
-          />
-        </v-card>
-      </div>
-
-      <div
-        v-else
-        :key="activity._id"
-      >
-        {{ $t("noDataAvailable") }}
-      </div>
-    </template>
-
-    <v-btn
-      color="primary"
-      fixed
-      bottom
-      left
-      @click="$router.go(-1)"
-    >
-      {{ $t("back") }}
-    </v-btn>
-  </div>
-</template>
-
-<style lang="scss" scoped>
-.container {
-  background: #efeff2;
-  height: 100%;
-  width: 100%;
-}
-
-.wrapper {
-  width: 90%;
-  max-width: 1024px;
-  margin: 2rem auto;
-}
-
-.v-card {
-  margin-bottom: 1.5rem;
-}
-
-.export-btn {
-  font-size: 0.8rem !important;
-}
-
-.export-btn .material-icons {
-  font-size: 1rem !important;
-  margin-right: 0.2rem;
-}
-
-.loading {
-  text-align: center;
-}
-
-.status-message {
-  margin-top: 40px;
-  padding: 20px;
-  border: 1px solid #ccc;
-}
-
-.chart-card {
-  padding: 1.5rem;
-}
-
-.activity-header {
-  margin: 1.5rem 0;
-}
-</style>
-
-<script>
-import _ from "lodash";
-import api from "../Components/Utils/api/api.vue";
-import Applet from "../models/Applet";
-import Activity from "../models/Activity";
-import Item from "../models/Item";
-import TokenChart from "../Components/Charts/TokenChart.vue";
-
-export default {
-  name: "TokenLoggerDashboard",
-
-  /**
-   * Components that this component depends on.
-   */
-  components: {
-    TokenChart,
-  },
-
-  /**
-   * Component state.
-   */
-  data: () => ({
-    loading: true,
-    status: "Initializing dashboard",
-    applet: {
-      _id: "",
-      label: { en: "Applet" },
-      activities: [],
-    },
-    charts: [],
-    responses: [],
-  }),
-
-  /**
-   * Computed properties.
-   */
-  computed: {
-    currentApplet() {
-      return this.$store.state.currentAppletData;
-    },
-  },
-
-  /**
-   * Handlers to be executed each time a property changes its value.
-   */
-  watch: {},
-
-  /**
-   * Method to be executed after the component has been mounted.
-   *
-   * @return {void}
-   */
-  async mounted() {
-    const { appletId } = this.$route.params;
-    const { users } = this.$route.query;
-
-    try {
-      this.status = "Loading applet data";
-      this.applet = await Applet.fetchById(
-        appletId,
-        {
-          withActivities: true,
-          withResponses: true,
-          withVersions: true,
-          users: Array.isArray(users) ? users : [users],
-        },
-        this.$store.state.currentAppletData.applet.encryption
-      );
-
-      this.loading = false;
-    } catch (error) {
-      this.status = "Oops, something went wrong. We couldn't load the applet";
-      this.error = error;
-      console.error(error);
-    }
-  },
-
-  /**
-   * Component methods.
-   */
-  methods: {
-
-  },
-};
-</script>
