+++ src/plugins/locales/en_US.js
@@ -41,7 +41,7 @@ export default {
   editors: 'Editors',
   manager: 'manager',
   managers: 'Managers',
-  owner: 'owner ',
+  owner: 'owner',
   institutionalID: 'Institutional ID',
   sendLinkPswdHosted: 'Send a link to reset your password hosted at ',
   loadingApplets: 'Loading Your Applets',
@@ -306,6 +306,7 @@ export default {
   resetSelection: 'Reset Selection',
   exportUserData: 'User Data Export',
   exportData: 'Export Data',
+  switchAccount: 'Switch Account',
   selectAppletToExportData: 'Please select applet to export data from: ',
   responses: 'Responses',
   tokens: 'Tokens',
